// src/modules/core/clusters/md-user/pages/user/types/ManageUserPage.types.ts
export type UserLogActivityParams = {
  user_id: string
  enabled: boolean
  page?: number
}

export type UserLog = {
  date: string
  data: TypeHistoricalData[]
}

export type UserLogActivityProps = {
  userLogs: UserLog[]
}

type Activity = {
  id: number
  action: 'create' | 'update' | 'login'
  timestamp: string
  user: {
    first_name: string
    last_name: string
    avatar: string | null
    job_function_id: {
      name: string
    }
  }
}

export type GetUserLogActivityResponseData = {
  id: number
  collection: string
  item: string
  delta: any
  activity: Activity
}

type Meta = {
  total_count: number
  filter_count: number
  current_page: number
  limit: number
  total_page: number
}

export type GetUserLogActivityQueryResponse = {
  data: GetUserLogActivityResponseData[]
  meta: Meta
}